change table products price column to NUMERIC

+++ migrations/sqls/20210803224008-products-table-up.sql
@@ -2,5 +2,5 @@
 CREATE TABLE products (
   id SERIAL PRIMARY KEY,
   name VARCHAR(64) NOT NULL,
-  price integer NOT NULL
+  price NUMERIC(9,2) NOT NULL
 );
